# backend/openapi.yaml
openapi: 3.0.3
info:
  title: OneSec API
  version: 1.0.0
  description: REST API магазина цифровых товаров OneSec (Express + Prisma)
servers:
  - url: http://localhost:4000/api
    description: Backend (dev)
  - url: http://localhost:3000/api
    description: Через Next.js rewrite
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /products:
    get:
      summary: Список товаров с фильтрами и пагинацией
      parameters:
        - name: page
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer }
        - name: q
          in: query
          schema: { type: string }
        - name: category
          in: query
          schema: { type: string }
        - name: discount
          in: query
          schema: { type: string, enum: ['1'] }
        - name: currency
          in: query
          schema: { type: string, enum: ['1'] }
        - name: inStock
          in: query
          schema: { type: string, enum: ['1'] }
      responses:
        '200':
          description: products + pagination
  /products/{id}:
    get:
      summary: Товар по ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: product
        '404':
          description: Не найден
...
  /admin/products:
    get:
      summary: Товары (admin)
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: products
    post:
      summary: Создать товар
      security: [{ bearerAuth: [] }]
      responses:
        '201':
          description: product
...